+++ packages/arango/data-loader/ArangoDataLoaderTypes.ts
@@ -8,14 +8,14 @@ import type {
   IfDbDoesNotExistOnGet,
   IfCollectionDoesNotExistOnGet,
   ArangoHostConfig,
-  CreateDatabaseUser
+  CreateDatabaseUserOptions
 } from '../utils/ArangoUtilsTypes'
 
 
 import { CollectionType } from '../utils'
 
-// re-export arang types
-export { CreateDatabaseUser }
+// re-export arango types
+export { CreateDatabaseUserOptions }
 
 export type IfTargetDbDoesNotExist = IfDbDoesNotExistOnGet
 export type IfTargetCollectionDoesNotExist = IfCollectionDoesNotExistOnGet
@@ -24,13 +24,16 @@ export interface LoadSpreadsheetDataOpts extends
   Pick<ParseOptions, 'reportProgress' | 'reportWarnings'> {
   ifTargetDbDoesNotExist?: IfTargetDbDoesNotExist
     // defaults to create
-  users?: CreateDatabaseUser[]
+  users?: CreateDatabaseUserOptions[]
     // only needed if IfTargertDbDoesNotExist is Create
     // defaults to []
   ifTargetCollectionDoesNotEist?: IfTargetCollectionDoesNotExist
     // defaults to Append
   validateEdgeTargets?: boolean
     // For edge collections, validate that the _from and _to docs exist
+    // defaults to true
+  validateGraphCollections?: boolean
+    // graphs, validate that the from and to collections exist
     // defaults to true
 }
 
@@ -67,4 +70,3 @@ export interface ArangoDataLoaderClientData {
 export const validWorksheetTypes = [
   'docCollection', 'edgeCollection','graph'
 ]
-