+++ packages/arango/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stcland/arango
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @stcland/spreadsheet-parser@0.0.4
+
+## 0.1.0
+
+### Minor Changes
+
+- Add support for parsing graph info from spreadsheet loader
+
 ## 0.0.3
 
 ### Patch Changes