+++ packages/spreadsheet-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stcland/spreadsheet-parser
 
+## 0.0.4
+
+### Patch Changes
+
+- Add support for colletion parsing
+
 ## 0.0.3
 
 ### Patch Changes